requirements.txt
nrequests
python-dotenv
beautifulsoup4
